changed the code so it will take the conditionals for the embed limits into account

+++ commands_to_load/Mod.py
@@ -14,10 +14,11 @@ class Mod():
         logger.info('[Mod rekt()] sending troll to unauthorized user')
         lol = '[secret](https://www.youtube.com/watch?v=dQw4w9WgXcQ)'
         eObj = em(title='Minion Things', author=settings.BOT_NAME, avatar=settings.BOT_AVATAR, description=lol)        
-        msg = await ctx.send(embed=eObj)
-        await asyncio.sleep(5)
-        await msg.delete()
-        logger.info('[Mod rekt()] troll message deleted')
+        if eObj is not False:
+            msg = await ctx.send(embed=eObj)
+            await asyncio.sleep(5)
+            await msg.delete()
+            logger.info('[Mod rekt()] troll message deleted')
     
     def __init__(self, bot):
         self.bot = bot
@@ -55,7 +56,8 @@ class Mod():
 
         name = ctx.author.nick or ctx.author.name
         eObj = em(description=desc, author=name, avatar=ctx.author.avatar_url, colour=0xffc61d ,content=fields)
-        await ctx.send(embed=eObj)
+        if eObj is not False:
+            await ctx.send(embed=eObj)
 
     @commands.command(aliases=['warn'])
     async def modspeak(self, ctx, *arg):
@@ -77,7 +79,8 @@ class Mod():
             msg += wrd + ' '
 
         eObj = em(title='A Bellow From the Underworld says...', colour=0xff0000, author=ctx.author.display_name, avatar=ctx.author.avatar_url, description=msg, footer='Moderator Warning')
-        await ctx.send(embed=eObj)
+        if eObj is not False:
+            await ctx.send(embed=eObj)
 
 def setup(bot):
     bot.add_cog(Mod(bot))
